Remove aboutlibraries dependency
The last version to support View is 11.5. It still works. But then they decided to deprecate it in favor of Jetpack Compose. There's no way I'm migrating to that any time soon, so fuck it, I'll just remove the library, it's useless. It hasn't been working for the past year and a half in release builds anyway, and nobody noticed.