More system UI misery fixes
The previous fix for deprecated statusBarColor didn't work at all for action mode. It created a blank status bar above that wouldn't disappear afterwards.

It used to work with a previous material library version... I don't know what changed so for now I'll leave it as is. The status bar is black during action, not a big deal.

Also:
- Fix navigation drawer left padding in landscape mode with navigation bar on the left.
- Use support action mode to get consistent action mode look across APIs.