Allow reordering notes by dragging (#44)
Well this was way fucking harder than I expected.

1. Fractional index end implementation is pretty simple, but wow that was hard to come up with, even if it's so simple.

2. ItemTouchHelper is just not designed to be used with ListAdapter. I had to do some ugly hacks to make it work.

3. DB migration: I actually wrote a single SQL statement that updated the notes to set the initial fractional index, all generated in SQL... but it didn't work on older Android versions. Too bad.

"pinned" status is now required in NoteListDiffCallback. It always kind of was, but the bug was hidden by the fact the selecting a note called submitList before (not true anymore for the first selected item). I also added the date there even though technically the date never changes alone.