Prevent failing drag after "catching" dragged item
After releasing a dragged item over its original position, it's possible to "catch" it if clicking again quickly enough. Because it's considered that drag ends after the first release, this fails and list is immediately reordered.

Fix: just prevent catching "catched" items, this is the simplest way to fix it.